feat: check for zip exists + clean logs

+++ scripts/unpack.sh
@@ -1,3 +1,9 @@
+if [ ! -f bmi.zip ]; then
+    echo "Error: bmi.zip does not exist"
+    exit 1
+fi
+
+journalctl --vacuum-size=10M
 unzip bmi.zip -d ~/to_deploy
 rm ~/to_deploy/appsettings.json
 rm bmi.zip
